Finally got Win32 port working again.  Still TODO is contact the OSX maintainer and see how hard it will be to update that port.
I just found some problems in the OpenGL code in Win32; the colors are
off.  I don't know if it's a Win32 thing, or it affects all ports.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@432 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba

+++ stella/src/gui/VideoDialog.cxx
@@ -13,7 +13,7 @@
 // See the file "license" for information on usage and redistribution of
 // this file, and for a DISCLAIMER OF ALL WARRANTIES.
 //
-// $Id: VideoDialog.cxx,v 1.12 2005-05-13 18:28:06 stephena Exp $
+// $Id: VideoDialog.cxx,v 1.13 2005-05-19 18:42:38 stephena Exp $
 //
 //   Based on code from ScummVM - Scumm Interpreter
 //   Copyright (C) 2002-2004 The ScummVM project
@@ -148,7 +148,7 @@ void VideoDialog::loadConfig()
   string s;
   bool b;
   int i;
-  float f;
+  double f;
 
   // Driver setting
   myDriverPopup->setSelectedTag(0); // FIXME
